feat: Separate webpack common and local.
